requirements_base_big.txt
alabaster==0.7.9
anaconda-clean==1.0
anaconda-client==1.5.1
anaconda-navigator==1.3.1
appdirs==1.4.3
argcomplete==1.0.0
asgi-redis==0.14.1
asgiref==0.14.0
astroid==1.4.7
astropy==1.2.1
attrs==16.3.0
autobahn==0.16.0
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.1
bitarray==0.8.1
blaze==0.10.1
bokeh==0.12.2
boto==2.42.0
Bottleneck==1.1.0
cached-property==1.3.0
cffi==1.7.0
channels==0.17.3
chest==0.2.3
click==6.6
cloudpickle==0.2.1
clyent==1.2.2
colorama==0.3.7
conda==4.3.14
conda-build==2.0.2
configobj==5.0.6
contextlib2==0.5.4
coverage==4.2
cryptography==1.5
cycler==0.10.0
Cython==0.24.1
cytoolz==0.8.0
daphne==0.14.3
dask==0.11.0
datashape==0.5.2
decorator==4.0.10
dill==0.2.5
dj-database-url==0.4.1
Django==1.8.8
django-countries==4.0
django-easymoney==0.7.1
django-floppyforms==1.7.0
django-idmap==0.4.1
django-vanilla-views==1.0.4
djangorestframework==3.4.6
docker==2.1.0
docker-compose==1.11.2
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
dynd==0.7.3.dev1
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.6
Flask==0.11.1
Flask-Cors==2.1.2
gevent==1.1.2
greenlet==0.4.10
h5py==2.6.0
HeapDict==1.0.0
honcho==0.7.1
huey==1.2.0
idna==2.1
imagesize==0.7.1
IPy==0.83
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
lazy-object-proxy==1.2.1
llvmlite==0.13.0
locket==0.2.0
lxml==3.6.4
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.3
mock==2.0.0
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.8
nb-anacondacloud==1.2.0
nb-conda==2.0.0
nb-conda-kernels==2.0.0
nbconvert==4.2.0
nbformat==4.1.0
nbpresent==3.0.2
networkx==1.11
nltk==3.2.1
nose==1.3.7
notebook==4.2.3
numba==0.28.1
numexpr==2.6.1
numpy==1.11.2
odo==0.5.0
openpyxl==2.3.2
otree-core==1.2.5
otree-save-the-change==1.1.3
packaging==16.8
pandas==0.18.1
partd==0.3.6
path.py==0.0.0
pathlib2==2.1.0
patsy==0.4.1
pbr==1.10.0
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.7.4
Pillow==3.3.1
pkginfo==1.3.2
ply==3.9
prompt-toolkit==1.0.3
psutil==4.3.1
psycopg2==2.6.2
ptyprocess==0.5.1
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.3.0
Pygments==2.1.3
pylint==1.5.4
pyOpenSSL==16.2.0
pyparsing==2.2.0
pytest==2.9.2
pytest-django==3.0.0
python-dateutil==2.5.3
python-redis-lock==3.2.0
pytz==2016.6.1
PyYAML==3.12
pyzmq==15.4.0
QtAwesome==0.3.3
qtconsole==4.2.1
QtPy==1.1.2
raven==5.25.0
redis==2.10.5
requests==2.11.1
rope-py3k==0.9.4.post1
ruamel-yaml===-VERSION
schema==0.6.2
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.18.1
service-identity==16.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.4.6
spyder==3.0.0
SQLAlchemy==1.0.13
stats==0.1.2a0
statsmodels==0.6.1
sympy==1.0
tables==3.2.3.1
terminado==0.6
texttable==0.8.7
toolz==0.8.0
tornado==4.4.1
traitlets==4.3.0
Twisted==16.2.0
txaio==2.5.1
unicodecsv==0.14.1
wcwidth==0.1.7
websocket-client==0.40.0
Werkzeug==0.11.11
whitenoise==3.2.1
widgetsnbextension==1.2.6
wrapt==1.10.6
ws4py==0.3.5
xlrd==1.0.0
XlsxWriter==0.9.3
xlwt==1.1.2
zope.interface==4.2.0
